+++ src/main/java/org/ecommerce/analyticsservice/dtos/ProductWithOrderStatusRevDto.java
@@ -12,8 +12,8 @@ import java.math.BigDecimal;
 public final class ProductWithOrderStatusRevDto {
     private String productId;
     private String productName;
-    private BigDecimal completedTotalPrice;
-    private BigDecimal cancelledTotalPrice;
-    private BigDecimal processingTotalPrice;
-    private BigDecimal deliveredTotalPrice;
+    private double completedTotalPrice;
+    private double cancelledTotalPrice;
+    private double processingTotalPrice;
+    private double deliveredTotalPrice;
 }
